✅ Add Test for modified ProductForm
